jffs2reader: scan all inodes for file metadata and content
Scan all inodes in the image when showing image or file contents.
To get exact file sizes all versions of the inode must be scanned. The
same goes for extracting files from the image, all inodes must be
scanned in order and extracted.
It's not spectacularily efficient but does the job.

Signed-off-by: Valentin Manea <valentin.manea@gmail.com>